requirements.txt
Flask==2.1.3
gunicorn
Werkzeug==2.3.7
requests
pydub
transformers
keras==3.3.0
librosa
spotipy
python-dotenv
torch